move test util
